Add function to Create/Save/Load API context
Used to easily register the device and api key, so that the context can be saved to the file and be re-used in the future, via the LoadContext function.
Also added an option to set the permitted ip addresses. bunq.WildcardIP or bunq.CurrentIP can be used to respectively allow said ranges. Alternatively, a custom slice of strings containing specific ip addresse or ranges may also be used.